Implementation of Gap Aware
Planned to be used on the "almost-last" partition
(or others, when comparing to pipedream...)
Added trick to save memory compared to original implementation.

Gap calculation assumes staleness 1, so skipping first batch.

It should be used in cases were there are accuracy problems
(as it penalizes gradients and consumes more compute and memory)

Note it uses Adam bias correction so it accelerates initial learning.