Fixes to allow running wrn28x10. NonInplace ReLus.
There is a problem when an inplace op (like relu) is the first op in the partition (because we make the first tensor a leaf, and can't do inplace ops on leafs).
Partition and Task now support multiple values and unpacking (waiting for test results).
Added  parttitioned resnet model.
Added configs to run several experiments.

TODO: seems like It has bad balancing AND/OR communication bottleneck
(starting at partition 1).