Allow partitioning without the LM head. (set this as default)
Note:
In pipeline for LM, it make sense that the only the last parition will know about the label, so the "WithLMHead" classes from huggingface are not what we need.
However, the last linear layer **should** be considered when balancing because its computationally heavy, so need to find some solution.

A simple solution I think of is to artificially add the computation time of that linear layer to the last output node (which will be in the last partition anyway).